+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Next
- - Added support for a UINT64 attribute type.
+ - Added support for a UINT64 attribute type. (Closes #34)
  - Broadcast discovery for all devices, even if one in the middle fails.
  - Omit attributes from devices with unknown types.